refactor: NumericValidator를 util패키지로 이동
