+++ src/test/kotlin/LottoControllerTest.kt
@@ -1,0 +1,2 @@
+class LottoControllerTest {
+}
